Separate DataBinding concerns into its own aspect. Move ownership of event handler for size change entirely in WindowsXamlHostBase.
